netqual/requirements.txt
# NetQual — Applied Networking Quality Validator
# Install: pip install -r requirements.txt

# Testing
pytest>=7.4.0
pytest-timeout>=2.1.0
allure-pytest>=2.13.0

# Config loading (protocol_base.py reads netqual_config.yaml)
pyyaml>=6.0
